leanstreams always sends 8 bytes for the header
The message length header is an int64 and leanstreams was trying to
send as few of bytes possible for the header. This was causing us to
receive an EOF error any time we set the MaxMessageSize to a number
greater than 65535.

This commit changes leanstreams to always send 8 bytes for the int64
header.